Fix: move detectPlatform function to utils
